Move Guzzle HTTP Transport to a separate class. Keep backward compatibility.
