+++ src/Klarna/Rest/Transport/ResponseValidator.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright 2014 Klarna AB
+ * Copyright 2019 Klarna AB
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
